requirements.txt
gradio
google-generativeai
gtts
